- added capablity to only use RIP when supported
